Added more parameters

+++ Header/VoiceDefinitions.h
@@ -11,7 +11,7 @@ namespace VoiceDefinitions
 	[[nodiscard]] std::string_view GetFilterTypeFromByte(uint8_t b);
 	[[nodiscard]] uint16_t ConvertByteToFilterFrequency(uint8_t b);
 	[[nodiscard]] double ConvertByteToFineTune(int8_t b);
-	[[nodiscard]] double ConvertByteToFilterQ(uint8_t b);
+	[[nodiscard]] float ConvertByteToFilterQ(uint8_t b);
 
 	[[nodiscard]] constexpr float GetBottomSectionPercent(const uint8_t value)
 	{
